Location data class to translate JSON

+++ app/src/main/java/ch/ti8m/gol/bbw_route/domain/entity/Coordinate.kt
@@ -2,4 +2,4 @@ package ch.ti8m.gol.bbw_route.domain.entity
 
 import com.google.gson.annotations.SerializedName
 
-data class Coordinate(@SerializedName("type") val type: String, @SerializedName("x") val x: Long, @SerializedName("y") val y: Long)
+data class Coordinate(@SerializedName("type") val type: String, @SerializedName("x") val x: Double, @SerializedName("y") val y: Double)
